requirements.txt
asgiref==3.8.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-mssql==1.8
django-mssql-backend==2.8.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==22.0.0
mssql-django==1.5
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.8.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
whitenoise==6.7.0